Separate fever tracking elements out of FeverBar
